requirements.txt
# Python version specification (3.11+)
# Note: This is a comment, not a pip requirement

# Core FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM - Latest available SQLAlchemy 2.x
sqlalchemy==2.0.43
sqlmodel==0.0.14

# Template and static file handling
jinja2==3.1.2
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0

# Date handling
python-dateutil==2.8.2

# PostgreSQL adapter
psycopg2-binary==2.9.9

# PDF generation
reportlab==4.0.4
